[BEAM-1585] Add beam plugins as pipeline options
