+++ sdks/python/apache_beam/options/pipeline_options.py
@@ -554,6 +554,14 @@ class SetupOptions(PipelineOptions):
          'worker will install the resulting package before running any custom '
          'code.'))
     parser.add_argument(
+        '--beam_plugins',
+        default=None,
+        help=
+        ('Bootstrap the python process before executing any code by importing '
+         'all the plugins used in the pipeline. Please pass a comma separated'
+         'list of import paths to be included. This is currently an '
+         'experimental flag and provides no stability.'))
+    parser.add_argument(
         '--save_main_session',
         default=False,
         action='store_true',
